+++ Clients/TgDownloaderDesktop/ViewModels/TgContactsViewModel.cs
@@ -72,7 +72,7 @@ public sealed partial class TgContactsViewModel : TgPageViewModelBase
 	{
 		if (!SettingsService.IsExistsAppStorage)
 			return;
-		var storageResult = await ContactRepository.GetListDtoAsync(take: 0, skip: 0, isNoTracking: false);
+		var storageResult = await ContactRepository.GetListDtoAsync(take: 0, skip: 0, isReadOnly: false);
 		List<TgEfContactDto> sourcesDtos = storageResult.IsExists ? storageResult.Items.ToList() : [];
 		SetOrderData(sourcesDtos);
 	}